backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.3
python-dotenv>=1.0.0
numpy
pandas
scikit-learn
joblib
python-multipart

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2
